Bug fixes
- Fixed Polygonal2DSelection#getVolume arithmetic, now returns number of blocks whose centers are inside the polygon
- Made Selection#getVolume return long
- Polygonal2DSelection#contains was posing issues due to the use of Polygon, is not purely arithmetic-based. Use of Shapes in Selection will likely be phased out

+++ src/main/java/com/monst/bankingplugin/commands/bank/BankCommandExecutor.java
@@ -170,7 +170,7 @@ public class BankCommandExecutor implements CommandExecutor, Confirmable {
 			p.sendMessage(Messages.SELECTION_OVERLAPS_EXISTING);
 			return true;
 		}
-		int volume = selection.getVolume();
+		long volume = selection.getVolume();
 		int volumeLimit = bankUtils.getVolumeLimit(p);
 		if (!isAdminBank && volumeLimit != -1 && volume > volumeLimit) {
 			plugin.debug("Bank is too large (" + volume + " blocks, limit: " + volumeLimit + ")");
@@ -503,7 +503,7 @@ public class BankCommandExecutor implements CommandExecutor, Confirmable {
 			p.sendMessage(Messages.WORLD_DISABLED);
 			return true;
 		}
-		int volume = selection.getVolume();
+		long volume = selection.getVolume();
 		int volumeLimit = bankUtils.getVolumeLimit(p);
 		if (bank.isPlayerBank() && volumeLimit != -1 && volume > volumeLimit) {
 			plugin.debug("Bank is too large (" + volume + " blocks, limit: " + volumeLimit + ")");
